+++ src/game/game.constants.js
@@ -2,3 +2,4 @@ export class Game {
 }
 Game.BLACK_JACK = 21;
 Game.DEALER_HIT_LIMIT = 17;
+Game.ODDS = 1.5;