+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 23/07/2025
+
+### Modified
+- Alerts instead console logs for debugging.
+- Role inheritance for player and dealer.
+
+### Added
+- Hidden card for dealer starting hand.
+- Flip hidden card to see it's value.
+- Player bet action.
+
 ## [0.0.1] - 21/07/2025
 
 ### Added